@@ -1,8 +1,8 @@
-import { IsTenantAvailableOutputState } from '@shared/service-proxies/service-proxies';
+import { TenantAvailabilityState } from '@shared/service-proxies/service-proxies';
 
 
 export class AppTenantAvailabilityState {
-    static Available: number = IsTenantAvailableOutputState._1;
-    static InActive: number = IsTenantAvailableOutputState._2;
-    static NotFound: number = IsTenantAvailableOutputState._3;
+    static Available: number = TenantAvailabilityState._1;
+    static InActive: number = TenantAvailabilityState._2;
+    static NotFound: number = TenantAvailabilityState._3;
 }
