@@ -4,6 +4,7 @@ using Microsoft.AspNetCore.Identity;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using Abp.Authorization;
+using Abp.Authorization.Users;
 using Abp.Configuration;
 using Abp.Domain.Uow;
 using AbpCompanyName.AbpProjectName.Authorization.Roles;
@@ -15,23 +16,16 @@ namespace AbpCompanyName.AbpProjectName.Identity
     public class SignInManager : AbpSignInManager<Tenant, Role, User>
     {
         public SignInManager(
-            UserManager userManager, 
+            UserManager userManager,
             IHttpContextAccessor contextAccessor,
-            UserClaimsPrincipalFactory claimsFactory, 
-            IOptions<IdentityOptions> optionsAccessor, 
+            UserClaimsPrincipalFactory claimsFactory,
+            IOptions<IdentityOptions> optionsAccessor,
             ILogger<SignInManager<User>> logger,
             IUnitOfWorkManager unitOfWorkManager,
             ISettingManager settingManager,
-            IAuthenticationSchemeProvider schemes) 
-            : base(
-                userManager, 
-                contextAccessor, 
-                claimsFactory, 
-                optionsAccessor, 
-                logger,
-                unitOfWorkManager,
-                settingManager,
-                schemes)
+            IAuthenticationSchemeProvider schemes,
+            IUserConfirmation<User> userConfirmation)
+            : base(userManager, contextAccessor, claimsFactory, optionsAccessor, logger, unitOfWorkManager, settingManager, schemes, userConfirmation)
         {
         }
     }
